Converted MetadataEditor to MetadataEditorContainer and added MetadataEditor component
